Move MultipleChecker into AnswerComposition namespace
As part of pulling our OpenAI integration out of the codebase, we simplified
a bunch of the code related to pipeline steps that were shared between
our OpenAI and Bedrock implementations.

As part of this work, the Jailbreak Guardrails were moved into a single
class in the answer composition pipeline directory.

This has left us with the MultipleChecker being the only class in the
Guardrails namespace.

Additionally, when I merged the MultipleChecker classes it was a bit of
a painful process due to the class being quite large with multiple
nested classes.

This commit breaks the MultipleChecker out into 3 classes, one handles
prompt constuction, another defined the response error and the final
class handles calling the LLM and parsing the response into a Result
object.

It then moves it into a MultipleGuardrail dirctory within the AnswerComposition
namespace.
