+++ lib/answer_composition/pipeline/answer_guardrails.rb
@@ -17,7 +17,7 @@ module AnswerComposition
           context.answer.assign_attributes(answer_guardrails_status: :pass)
           context.answer.assign_metrics(guardrail_name, build_metrics(start_time, response))
         end
-      rescue ::Guardrails::MultipleChecker::ResponseError => e
+      rescue MultipleGuardrail::ResponseError => e
         abort_after_response_error(context, e, start_time, Answer::CannedResponses::ANSWER_GUARDRAILS_FAILED_MESSAGE)
       end
     end
