+++ lib/answer_composition/pipeline/question_routing_guardrails.rb
@@ -19,7 +19,7 @@ module AnswerComposition
           question_routing_guardrails_status: response.triggered ? :fail : :pass,
           metrics: { guardrail_name => build_metrics(start_time, response) },
         )
-      rescue ::Guardrails::MultipleChecker::ResponseError => e
+      rescue MultipleGuardrail::ResponseError => e
         abort_after_response_error(context, e, start_time, Answer::CannedResponses::QUESTION_ROUTING_GUARDRAILS_FAILED_MESSAGE)
       end
     end
